Address review: fix the ternary, parse TOML, count base dependencies
