+++ changelog_entries/2049.improvement.md
@@ -1,1 +1,1 @@
-The formatwise gallery job now tests only the gallery pages a pull request changed, and a `pyproject.toml` edit triggers it only when `[project.optional-dependencies]` changed, instead of sweeping all 69 pages whenever the file was touched at all.
+The formatwise gallery job now tests only the gallery pages a pull request changed, and a `pyproject.toml` edit triggers it only when `[project.dependencies]` or `[project.optional-dependencies]` changed, instead of sweeping all 69 pages whenever the file was touched at all.
